+++ packages/materials/src/element/tag.ts
@@ -1,4 +1,5 @@
 import type { MaterialDescription } from '@vtj/core';
+import { effect, size, type } from '../shared';
 const Tag: MaterialDescription[] = [
   {
     name: 'ElTag',
@@ -9,46 +10,42 @@ const Tag: MaterialDescription[] = [
     package: 'element-plus',
     props: [
       {
-        name: 'type',
+        ...type('type'),
         defaultValue: 'primary',
-        options: ['primary', 'success', 'info', 'warning', 'danger'],
-        setters: 'SelectSetter'
       },
       {
         name: 'closable',
+        title: '是否可关闭',
         defaultValue: false,
         setters: 'BooleanSetter'
       },
       {
         name: 'disable-transitions',
         defaultValue: false,
-        label: '渐变动画',
+        label: '禁用渐变动画',
+        title: '是否禁用渐变动画',
         setters: 'BooleanSetter'
       },
       {
         name: 'hit',
+        title: '是否有边框描边',
         defaultValue: false,
         setters: 'BooleanSetter'
       },
       {
         name: 'color',
+        title: '背景色',
         defaultValue: '',
         setters: 'ColorSetter'
       },
+      size('size'),
       {
-        name: 'size',
-        defaultValue: 'default',
-        options: ['large', 'default', 'small'],
-        setters: 'SelectSetter'
-      },
-      {
-        name: 'effect',
-        defaultValue: 'light',
+        ...effect('effect'),
         options: ['dark', 'light', 'plain'],
-        setters: 'SelectSetter'
       },
       {
         name: 'round',
+        title: 'Tag 是否为圆形',
         defaultValue: false,
         setters: 'BooleanSetter'
       }
@@ -68,20 +65,18 @@ const Tag: MaterialDescription[] = [
     props: [
       {
         name: 'checked',
+        title: '是否选中',
         defaultValue: false,
         setters: 'BooleanSetter'
       },
       {
         name: 'disabled',
+        title: '是否禁用',
         defaultValue: false,
         setters: 'BooleanSetter'
       },
       {
-        name: 'type',
-        label: 'type',
-        title: 'CheckTag 类型',
-        setters: 'SelectSetter',
-        options: ['primary', 'success', 'info', 'warning', 'danger'],
+        ...type('type'),
         defaultValue: 'primary'
       }
     ],