+++ packages/materials/src/shared/props.ts
@@ -3,6 +3,7 @@ import type { MaterialProp } from '@vtj/core';
 export function size(name: string = 'size'): MaterialProp {
   return {
     name,
+    title: '尺寸',
     defaultValue: 'default',
     setters: 'SelectSetter',
     options: ['default', 'large', 'small']
@@ -12,8 +13,19 @@ export function size(name: string = 'size'): MaterialProp {
 export function type(name: string = 'type'): MaterialProp {
   return {
     name,
+    title: '类型',
     defaultValue: 'default',
     setters: 'SelectSetter',
     options: ['default', 'primary', 'success', 'warning', 'danger', 'info']
+  };
+}
+
+export function effect(name: string = 'effect'): MaterialProp {
+  return {
+    name,
+    title: '主题',
+    defaultValue: 'light',
+    setters: 'SelectSetter',
+    options: ['dark', 'light'],
   };
 }