+++ apps/app/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.2](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.15.1...vtj-project-app@0.15.2) (2026-02-06)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.15.1](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.15.0...vtj-project-app@0.15.1) (2026-01-28)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+# [0.15.0](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.6...vtj-project-app@0.15.0) (2026-01-27)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.14.6](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.5...vtj-project-app@0.14.6) (2026-01-26)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.14.5](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.4...vtj-project-app@0.14.5) (2026-01-24)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.14.4](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.3...vtj-project-app@0.14.4) (2026-01-16)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.14.3](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.2...vtj-project-app@0.14.3) (2026-01-09)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.14.2](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.1...vtj-project-app@0.14.2) (2026-01-06)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
 ## [0.14.1](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.14.0...vtj-project-app@0.14.1) (2026-01-04)
 
 **Note:** Version bump only for package vtj-project-app