+++ docs/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.2](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.15.1...@vtj/docs@0.15.2) (2026-02-06)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.15.1](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.15.0...@vtj/docs@0.15.1) (2026-01-28)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+# [0.15.0](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.6...@vtj/docs@0.15.0) (2026-01-27)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.14.6](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.5...@vtj/docs@0.14.6) (2026-01-26)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.14.5](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.4...@vtj/docs@0.14.5) (2026-01-24)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.14.4](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.3...@vtj/docs@0.14.4) (2026-01-16)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.14.3](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.2...@vtj/docs@0.14.3) (2026-01-09)
+
+
+### Bug Fixes
+
+* 🐛 修复代码不完整，似乎被截断了的提示 ([c45b5b2](https://github.com/samchen08/vtj.pro/commit/c45b5b291facee570a7a90c22cd6d2559adedff5))
+
+
+
+
+
+## [0.14.2](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.1...@vtj/docs@0.14.2) (2026-01-06)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
 ## [0.14.1](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.14.0...@vtj/docs@0.14.1) (2026-01-04)
 
 **Note:** Version bump only for package @vtj/docs