+++ packages/renderer/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.2](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.15.1...@vtj/renderer@0.15.2) (2026-02-06)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
+## [0.15.1](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.15.0...@vtj/renderer@0.15.1) (2026-01-28)
+
+
+### Bug Fixes
+
+* 🐛 node ref 在mounted后获取不到问题 ([b3a33f5](https://gitee.com/newgateway/vtj/commits/b3a33f5d842e7bf343929a2b5fa7fccf2d9bda0a))
+
+
+
+
+
+# [0.15.0](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.6...@vtj/renderer@0.15.0) (2026-01-27)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
+## [0.14.6](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.5...@vtj/renderer@0.14.6) (2026-01-26)
+
+
+### Bug Fixes
+
+* 🐛 修复区块组件缓存问题 ([6601a00](https://gitee.com/newgateway/vtj/commits/6601a0083249a2490e12f1ae685f1d0c6e841dd5))
+
+
+
+
+
+## [0.14.5](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.4...@vtj/renderer@0.14.5) (2026-01-24)
+
+
+### Bug Fixes
+
+* 🐛 修复在线异步区块渲染顺序问题 ([e758023](https://gitee.com/newgateway/vtj/commits/e75802320b96f8ec9f766bd419c27213ed921464))
+
+
+
+
+
+## [0.14.4](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.3...@vtj/renderer@0.14.4) (2026-01-16)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
+## [0.14.3](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.2...@vtj/renderer@0.14.3) (2026-01-09)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
+## [0.14.2](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.1...@vtj/renderer@0.14.2) (2026-01-06)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
 ## [0.14.1](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.14.0...@vtj/renderer@0.14.1) (2026-01-04)
 
 **Note:** Version bump only for package @vtj/renderer